refactor: Extract hash approval and watcher logic into dedicated services
